"2:45pm. Done with breakfast. Let me just read the OPM chapter and then I will look into how to put illustrations into the chapter.
3:25pm. I had my fun. Let me get some work done. Let me figure out how to post illustrations.

3:30pm. Oh, I see that the add chapter has a lot more options. There are pre-chapter and post-chapter notes. Also there is something called clean paste specifically for pasting from Google Docs.

In insert image, is it asking me for an URL?

https://www.royalroad.com/forums/thread/101959?page=1#pid839802
> To go a bit more in-depth: You'll need a place online that hosts the image. For this purpose, you either have a webspace of your own or you use one of the many image hosters out there. One popular choice is https://imgur.com/. It's mostly reliable and doesn't do any nonsense to your image. For the rest of the explanation, I'm going to base it of Imgur.

> What you do is you upload the image by clicking new post on imgur. You point it to your file and it will upload. After that, you get a window with the image in the website and various options. When you point at the image (equation in your case), you'll get a window popup that says copy. You click on the v and pick "Get share links". There you pick the option BBCode Forums, Copy. It looks something like this

I wonder if I could use Github or Twitter instead to host my images?

3:35pm. It is really annoying how auto-registering with Google is not working for me.

https://imgur.com/a/blBYUA2

Chapter must be at least 500 characters long.

Ah for fuck's sake.

I guess I need to link directly to the image.

3:50pm. Now RR is messing with the proportions.

I need to scale them down to 1280 otherwise RR will squeeze it down badly.

3:55pm. Let me put in Euclid's Room illustrations.

> First, Thank you for your work.
> It seems that style weight added to your implementation. But why is there no style weight in the option?

I didn't see this tweet earlier. But I am not sure...

4:35pm. Did some replying.

https://www.royalroad.com/fiction/57747/simulacrum-heavens-key/chapter/973184/the-first-dream

Anyway here is the post. What I need to do next is post the rest of the chapters. Let me actually post just one. I want to proofread them properly, and am feeling tired right now.

4:45pm. I actually didn't proofread the first chapter. I should probably go back and do it. But not right now. Let me just deal with two, post it and then I will start looking into opening a Patreon account.

4:55pm. Found a bunch of errors in chapter 2.

5pm. Ah, right. I should not forget to post it on Twitter.

I am almost done with proofreading.

5:05pm.

> (Bad End - The First Nightmare)

Rather than call it a nightmare, how about...

> (Bad End - The Dream Left Behind)

...No, let me stick with the original. It is not that important what the first bad end's name is. I won't obsess over it.

///

Imgur Album: https://imgur.com/a/blBYUA2

I modeled the room and the VN cover in Blender. The reference for the room is my own. I could have taken a photograph, but instead I spent an insane amount of time modeling the objects in Moi 3d and importing them in Blender. Altogether, from late 2021 till now, I spent 9 months studying art, but I hadn't gotten to the point where I could make illustrations quickly enough to make it viable to make use of it for my own novels. Maybe if I spent another few years practicing, I could get to that level, but instead I threw in the towel and decided to just write instead. I really just want to write instead of spending 2-3 weeks per illustration. Art is quite hard, and my talent in it is mediocre, so it has been quite a slog to even get to this point.

This is not the end of the story though when it comes to graphics. At this point in time, generative NNs have gotten really good, and I plan to make use of Stable Diffusion when it comes out of beta as my illustrative tool.

I'll probably get access to it before I manage to get approved on DALLE's waiting list.

My GTX 970 won't be powerful enough to run Stable Diffusion, so I'll have to take some time to set it up on the cloud. I've read that it should come out of beta next week, but I am fine whenever.

It really seems fitting that a story about the Singularity would be illustrated by a NN.

For now my plan is to just write, I'll come back to the early chapters and add more imagery once I get access to a generative NN on par with DALLE. On Royal Road, I've set the image resolution to 1280 x 720 to prevent it from squeezing them, but you open them in a separate window, you'll be able to see them in their full glory.

In addition to a generative NN to create whole images, I'll be using style transfer on top of that as well. I've mentioned that I've modeled the above in Blender. Afterwards I did a style transfer step to add interest.

I've used this tool to do it: https://twitter.com/Ghostlike/status/1543600747568308226

You can see some example of style transfer in action on my profile.

///

///

Heaven's Key is supposed to be a VN, but it makes more sense for me to gradually work on it here on RR rather than in complete silence until I am done. It is also supposed to have music. I have a few good tracks in my head, but unlike art skills which I have 9 months invested in it, or programming which has 6.5 years worth of exp points, my music skills have literally 0, so that is going to have to wait.

Heaven's Key is also supposed to be anime, but char models, as opposed to background illustrations are going to have to wait as well. In terms of art skills my strength is that I am familiar with wide variety of art programs: Blender, Moi, Substance Painter and Designer, Houdini, Clarisse, Zbrush. I treated my art journey like a programming one, so although I have a decent amount of knowledge to base further learning on, to actually get better I'd need to sit down and actually start making pieces. My training as a programmer is working against me here. Programming is all about automation, so given a particular task I just end up looking for shortcuts, but after a certain point in order to get good at art you simply have to sit down and make things.

After 9 months I had to make a choice whether I want to spend more time studying art, or whether I should just write.

The fact that DALLE came out factored into that decision. I fought against the decision to just start writing for a while, but then got tired of 3d. It might turn out that in the future we'll get a generative NN model that is good at the anime aesthetic.

Looking into the future, it feels like being a writer is going to get all kinds of advantages that were originally only accessible to other professions. Throughout the last decade, I really expected programming itself to be the biggest beneficiary of the AI wave, but I was wrong about that. Right now the programming skill is not a big deal, but in the post-Singularity era it will be.

///

5:35pm. Did the second post. I won't bother posting the third today.

One thing I've realized is that maybe I should be adding the chapter numbers to each chapter. I'll do that later.

Right now, let me take a look at Patreon.

5:45pm. It turns out I already have a Patreon account. I must have done that while setting up sponsorship for Spiral.

'What are creating? A programming language.'

Yeah. It seems that I did not finish the page back then. I'll take the cover photo from my AngelList page. Or maybe I should put in some kind cartoon. I am not sure.

5:50pm. Is the cover image for my Patreon page or something else?

Let me stop here for the day, I am too tired to deal with this anymore.

5:55pm. Let me close here. Lunch then fun."