"1:35pm. After I am done with breakfast, I'll take a bath. I am overdue on that and I can't work in this filth. Afterwards...Yeah, I'll post the next chapter on RR. Actually, before that I have to figure out how to make the tiers. The 3E one will just be for sneak peeks. Initially I'll only have the 5 and 10E ones. 5 for all the chapters, and 10 for that plus this journal.
After that I'll go into the grove of writing two chapters and posting one on RR.

2:15pm. Done with breakfast. Let me watch ep 2 of Kunoichi Tsubaki and then I am going to start. I need to figure out how to get a handle on the monthly tiers.  Actually before I start, I need to take that bath. Sigh. Well, after I deal with today I'll be able to just get back to the writing groove.

2:45pm. It is a fun anime. Bath time. I've been looking forward to this.

3:25pm. Phew, I got that shave. Nice, I am set once again.

Ok, let me check out my Patreon page. Can I add a 5E tier?

Yeah, I just have to click the Edit Tiers button.

Right now I have the 5E/month Aleator and 10E/month Episteme tiers. I'll add the Introspect 3E/month tier later.

3:30pm. Now it is time to publish another chapter. I also want to go over the first one for any errors.

Before that, are the new 40xx cards out. I want to check out their power consumption.

https://www.gpumag.com/nvidia-geforce-rtx-4000-series/

No, not really. I won't bother reading this.

https://semianalysis.substack.com/p/intel-cuts-fab-buildout-by-4b-to
> First Net Loss In Over 30 Years, Cutting Fab Buildouts, But “Committed To Growing The Dividend”

I've been thinking about this yesterday, but I think Intel deserves some kind of disapointment of the decade award. Intel had everything going for it to lead the way to a multicore revolution. Had it done so, we could have had 1,000 core chips today instead of waiting for AI startups to get to them. If the Singularity misses its 2045 target, it is Intel, not memristors, to whom the blame should be pointed at.

https://semianalysis.substack.com/p/cxl-deep-dive-future-of-composable

Right now I am reading this. Let me finish it and then I will publish the story.

> Datacenters have a massive memory problem. Core counts have grown rapidly since 2012, but the memory bandwidth per core and capacity has not increased commensurately.

3:50pm. Ok, let me go over chapter 1. Then I'll go over chapter 3.

> simulacrum_post_singularity_story

Let me see if I can change the page name to this.

https://www.patreon.com/simulacrum_post_singularity_story

It works.

3:55pm. Hmmm, let me go with that as my patreon page. It is long, but rememberable.

https://boards.4chan.org/h/thread/6801965
AI generated hentai

Not too bad. A post says they will release Stable Diffusion on Monday.

Let me get back to chapter 1.

4:15pm. I am not really focusing on it too intently, but chapter 1 seems fine.

4:35pm. I didn't find a single line to edit in chapter 3. I guess it is fine.

Ah, before that, let me edit the chapter numbers. How do other people label them?

Something like 'Chapter 1: The First Dream' and so on is fine.

4:50pm. Let me write the post-chapter autor note.

///

8/19/2022

Patreon: https://www.patreon.com/simulacrum_post_singularity_story

No benefits yet, but I'd appreciate any support you could give. The plan right now is to just write a lot to prove my willingness to do this, and once I have sufficient volume out, I'll start leaving some aside for Patrons.

All the arcs of Simulacrum are less stories for me and more like my aspiration, both to enter the self improvement loop and what the world I would create would be like. I am actually quite embarrassed that I've turned to writing instead of pursuing the Singularity through my programming, but the path of transcendence is a very hard slog in today's world. I'd end up homeless on the street if I pursued it any harder. The defining feature of any path is that it gives you a way of starting off and then gradually multiplying your advantage. It is how investors compound their investments over time, or salarymen get promotions at their company. The way I've imagined the path of transcendence in this age would go is that I'd make a weak AI agent, and the use of it would give me the resources that I'd need to make better agents, but even something like making an agent for a toy game like poker is quite hard. I mean, it is not that trivial - from a certain perspective Go is also trivial, but AlphaGo needed a cluster to be trained on, and poker is more complex than it. Still, I had a certain vision and a belief in how things should go so I pursued it. The AI agent should have given me resources, and on my own end, I had created a programming language that would make it easy to program the AI chips. Spiral is quite a good functional programming language, you can get it as a VS Code extension if you are interested. I am proud of it. Where I fell short is the making of an AI agent. Since they had the greatest potential, I made the mistake of focusing on NNs and deep RL.

I couldn't really avoid making such a mistake as I thought a language that can be used to easily make ML libraries on AI chips would be necessary, so I had to focus on that. But given how poor deep RL methods are, even making a weak poker agent turns out to be too hard on a single GPU. There is an alternative and more natural path to AI, where you start off with handwritten rules - the good, old-fashioned AI, then bring in evolutionary methods to optimize whatever handwritten rules you have. That would bring you to the weak level without much issue. After that it would be possible to bring in NNs on top of evolutionary methods. This would be a much stabler path of development and it is a pity I haven't taken it.

It would not be that easy to bring in NNs into the mix with today's hardware though. GPUs really shine when you have static datasets, but when you have simulators or games, and want to pass the inputs as they come, that is without batching, you run into huge efficiency issues. If I could get access to AI chips, I could resolve that. Since I have good PL skills owing to me working for 3 years on my own language, I really should be making all kinds of tooling for AI chips, but it is hard to get in touch with these companies or get access to their chips.

///

5:35pm. I am going to close here for the day. I'll leave proofing the chapter 4, and writing the note for it for tomorrow.

I want to officially give Intel the disappointment of the decade award.

5:40pm. Oh right, let me add an image to the album.

Simulacrum: Heaven's Key Illustrations: https://imgur.com/a/blBYUA2

I'll use it as a VN scene at some point. I've already posted it in plenty of styles on my Twitter.

Ah, right, let me post a link to Heaven's Key on my Tweeter.

https://twitter.com/Ghostlike

Done.

5:45pm. Yeah, this is good enough. I'll really close for the day here. Tomorrow I'll post chapter 4, and then get to work on adding more to chapter 5.

Time for lunch."