"9:55am. I went to bed too late and now I am groggy.
>Starter guides:
https://rentry.org/retardsguide - Main txt2img guide
https://rentry.org/kretard - K-Diffusion guide
https://rentry.org/img2img - img2img guide

This is from the /g/ thread. The SD v1.4 weights are out, and I should get them. Since there is no way I can set this up on my own rig, I'll have to go the cloud route. Mggghhhh...this is quite a pain in the ass compared to just getting web UI like with DALLE.

10am. Doing my own UI, and setting all this up will take me a while. Compared to just running it on my own rig, cloud will be more complicated. Plus, it is not like I have actual cloud experience, doing it for the first time will make it take even longer than it should.

10am. For now, let me just set the weights to download.

The checkpoint file is 4gb. Ok.

10:05am. Right now let me just chill. I'll split studying the cloud and writing until everything is set up. It could very well take me a few weeks until that is all done. I won't be running this any time soon, but I am not in a hurry.

https://huggingface.co/spaces/stabilityai/stable-diffusion

Or maybe I could just run it from here.

https://twitter.com/EMostaque/status/1561777122082824192

They have something called dream studio.

https://stability.ai/blog/stable-diffusion-public-release

This post has a bunch of links.

///

DreamStudio is a front end and API to use the recently released stable diffusion image generation model.

It understands the relationships between words to create high quality images in seconds of anything you can imagine. Just type in a word prompt and hit “Dream”.

The model is best at one or two concepts as the first of its kind, particularly artistic or product photos, but feel free to experiment with your complementary credits. These correlate to compute directly and more may be purchased. You will be able to use the API features to drive your own applications in due course.

Adjusting the number of steps or resolution will use significantly more compute, so we recommend this only for experts.

The safety filter is activated by default as this model is still very raw and broadly trained. This may result in some outputs being blurred, please adjust the seed or rerun the prompt to receive normal output. This will be adjusted as we get more data. Please use this in an ethical, moral and legal manner. Make your mothers proud.

We hope you enjoy having a trillion images in your pocket, please do share and tag your output with #StableDiffusion!

///

10:55am. Had to do some chores. Anyway, I am finally ready to start.

11am. First of all price. I've checked out the Dream Studio. It is cheaper than DALLE, but at most something like 3x. It is not hugely so. I have the same amount of free generations as DALLE.

> Black fog covers the a skeleton who howls and grasps at the skies, arms raised, as he kneels facing upwards on the dark, densely foggy streets of a dilapidated, post-apocalyptic city, digital art.

I tried putting in this kind of prompt, but although the skeleton is of higher quality than with DALLE, it is not doing the pose correctly.

Let me try the prompt that got banned at DALLE.

> Howling skeletons in the darkness of a dilapidated city as the deathly, black fog devours the flesh from their bodies. Digital art.

It works, but it does not give me a human skeleton, but some sick abomination.

11:10am. > Black fog covers the an ivory boned skeleton who howls and grasps at the skies, arms raised, as he kneels facing upwards on the dark, densely foggy streets of a dilapidated, post-apocalyptic city. Digital art, trending on art station, detailed.

Let me save this here.

> Black fog covers the an ivory boned skeleton who howls and grasps at the skies, arms raised, as he kneels facing upwards on the dark, densely foggy streets of a dilapidated, post-apocalyptic city. The deadly nanomachine fog surrounds him, devouring the flesh from his body. Trending on art station, digital paint, 4k.

I think that increasing the Cfg did something to make the poses better. Let me try this.

Tsk, I keep getting blurry images. This is censorship.

11:20am. I do not think it is possible to turn it off in Dream Studio. I want to ask /g/ how to improve my prompts, but for some reason no SD thread is up.

11:25am. Goddamit.

Ok, let me read the guide.

https://rentry.org/retardsguide

So for the plan for this would be to take this, and run it to generate a smallish image of something like 32 x 32 on the CPU. I definitely won't be able to do deal with this on the GPU.

After that, I can start thinking about setting it up on the cloud. Honestly, this is kind of a pain in the ass.

Let me try one of the golden city prompts.

11:40am. It is going to take effort to beat this into shape.

Let me do it like this - I won't do the illustrations right now. I do not feel like messing with setting up the cloud when I should just be writing.

I can finish writing Heaven's Key and Hatred, and then come back to illustrate this stuff. At that point setting things up on the cloud would be a good programming exercise.

Right now, I want to avoid distruptions to my writing pace.

11:45am. I am already running into difficulties trying to get the image the way I want it. SD is really good at the skeletal model, but doesn't get the pose right. DALLE is the opposite. I have better things to do that twiddle my thumbs trying to get it to work.

Forget this, I was hyped for it, but now I am done.

Right now what I really want to do is just write, so that is what I should do. Forget time wasting obligations. Has any step trying to punch upward ever worked for me?

Also the images are 512 x 512. This is not high enough resolution. But if I ramp it up to 1k on SD, the cost per image goes up by 10x on the Dream Studio site. At that cost I only have room for generating 18 images.

11:50am. Years down the road there will be better and cheaper models. The time I take messing with these right now, I could be spending writing Heaven's Key. The tradeoff is not worth it.

So I'll resume writing. Let me have breakfast here."