"6:30pm. Oh shit, I was in bed for a lot longer than I thought. I couldn't help it. I am very sad today. As I think about the self improvement loop it just bring my own failures to bear. While doing 3d art I was able to avoid it for a while, but that damn failure essentially ruined me.
It is going to take me effort to overcome.

I have 3 goals.

1) Write Simulacrum.
2) The second one I will save for the author's note.
3) Try out the evolutionary approach. Put my effort where my mouth is and try it out on a small scale.

I can't really just not program for the next few years. Taking on a project to redeem myself is something I should challenge myself with.

Taking a meta approach to ML is something that will help me overcome my defeat.

It will help me move from putting my hope on the ML community or the hardware guys, to my own programming. Why not make that basic bitch poker agent using evo methods?

It is true that the true power of evolution lies in evolving learning algorithms, but I won't be able to try that path until the hardware gets good enough. But nonetheless, I know that this is what I will be doing.

If somebody gives me the brain core, and the human intelligence algorithms, how would I go to a level higher than that?

Of course, by using the power of evolution. With the brain core, that approach would be limitlessly more powerful. It would work.

It wouldn't work on my current rig. But there is something to be said for getting some practice early on. It would at least prove that I am not a complete fuck up.

6:35pm. Also I think I figured out what the purpose of sleep is.

The strength of evolutionary algorithms, is that I could continually build and add parameters, and optimize them piecemeal. At the end of that I could have a large program, that has been gradually built up. What I could do then, is that and distil them into a NN.

Deep RL is super shitty, but distiling a policy would just be supervised learning and work well.

After that I could use that NN as just another component in the program and start building it up again using evo methods. This would avoid all the instability of RL and give me guaranteed improvements.

I think that the brain does something similar - during the day it accumulates memories and during sleep it distils them. Why not optimize directly?

Because actual RL is very unstable, but supervised learning is the opposite.

Tabular RL is a red herring, I won't find a use for it.

Evolutionary algorithms are really good at one thing - optimizing parameters regardless of their placement. I could even take that original network, copy it, stick extra layers in it, and optimize those piece by piece and afterwards distil it into the original.

There is a point to this - by adding memories instead of mutating them, you can guarantee never losing the original progress.

6:45pm. Sigh, I won't regret learning 3d, but the last year sucked. The parts that really stick out for me are the job applications. I've never been so humiliated in my life. It feels like in the last year I went against everyhing I stand for.

The me that I idolize should have firm beliefs and confidence in his skills, right now I have neither.

I will take on the project on brushing up on evolutionary methods using poker. But it won't be like last time. Since this will be small scale, my goals will be humbler. It will just be preparation for the future.

To begin with, nature probably started off with simple behavioral programs and then added memory as it went along. I doubt that bacteria needed to sleep. I should follow in its example and do it step by step.

7pm. I've falled into a trap set by people like Hinton. For some reason I've internalized that doing things by hand is cheating, when I should have well known that there is a way path from handwritten rules to general behavior.

After I do my small scale tests, I won't have good enough to make money in real world tables, but at least I will be able to call myself a proper scientist. Evolutionary methods are the automation of science.

7:05pm. Let me commit here. Tomorrow I am going to see if I can get some writing done. If I can't I'll spend the day in bed again until I muster up the will.

7:10pm. I don't think I can really be myself without programming. Even if it is for small wins, I should aim for them. I should save my grandest thoughts for my writing, and my humblest thoughts for my programming."