"9:40am. Yesterday it was getting colder, but it is freezing right now in my room. I've had to bring in a heater.
Anyway, I've been lounging in bed for a bit too long today. Today I'll take it easy. First of all, let me chill, then I will do some test posts on Patreon.

It seems I've reached level 3 on Royal Road.

...Only a single view on what I've posted yesterday. It really doesn't bode well at all for my writing career. Maybe it is hopeless, but I'll execute the plan and see where it leads me.

9:45am. Hmmm, this is a problem. When I log into Patreon it puts me on the page I opened. Was that Patreon page supposed to be a central page for all my work? Can't I have multiple Patreon pages?

https://www.reddit.com/r/patreon/comments/4x6nft/patreon_and_multiple_pages_with_one_account/

9:55am. What a pain in the ass. Nevermind that for now, let me just chill a little. Then I will post all the chapter 8 pieces on Patreon.

10:35am. Let me start. I had time to think about what I want to do. First of all, let me post chapter 8 on the Patreon page.

I'll create two accounts, with 5/month subscriptions. I've imagined linking the accounts, but nevermind that for now.

https://www.patreon.com/simulacrum_heavens_key

Here is the Patreon page, I've changed its name. I'll have to update the two links later. For now, let me just figure out how to post things on Patreon.

10:40am. Sigh, Patreon is butchering my formatting!

10:45am. https://www.reddit.com/r/patreon/comments/vdt3pk/copying_from_any_text_editor_into_patreon_and/

This is so annoying.

10:55am. Sigh. So far the only way to post text properly on Patreon is to first post it into RR and the copy paste it to Patreon.

https://www.royalroad.com/forums/thread/122239

Kek, he is doing better than me.

///

Yes. Views aren't indicative of anything.

Views basically mean how many people clicked when your chapter appeared on the latest updates.

It's entirely random numbers that vary even if you have updates scheduled to post automatically, since you can't predict how many people are looking at the page at given time. While the first chapter has most views since it is the first thing everyone sees, rest is pretty much random, depending on unpredictable variables and circumstances saying nothing about the popularity of your story. You can actually guess which cover attracts most views, assuming you change covers regularly, but even then it is a stretch.

What is indicative is how many followers you gain every update. Followers are people who showed interest in your story, even if temporarily.

///

Hmmm, interesting.

11:15am. https://www.royalroad.com/forums/thread/122332

///

I've written a fair bit of my own story, and am trying to post it on Patreon.

When I copy paste it there, it looks a bit like on RR without clean paste enabled. The way I do on RR is that I right click and copy it as plain text. On RR it just works, but on Patreon I get ugly gaps between paragraphs. It works a bit better when I copy it to RR and then from there to Patreon as RR eliminates superfluous gaps, but I sometimes lose formatting on indented text.

This is really annoying. How do you pros do it?

Shouldn't there be a clean way of doing this? Is there a better site similar to Patreon which allow me to properly paste text?

///

I've opened this thread. You'd think that just posting some text would be easy, but it is anything but that.

11:30am. https://www.disciplemedia.com/community-monetizations/patreon-alternatives/

https://ko-fi.com/

How could the platform even work if it has no fees?

...Ah, whatever. Let me start a page.

> Connect to PayPal

Ugh.

11:35am. https://youtu.be/34ltAowKUvU
What to Know Before Starting Patreon

This is so annoying. I just want to paste text, and it is going to waste my whole day at this rate.

The way things are going, even my journal is going to end up having to be uploaded as a pdf.

https://www.patreon.com/posts/new-feature-29012102

It has markdown?

11:45am. This is going to be annoying.

https://www.patreon.com/posts/using-markdown-29519541

This probably has what I want, but of course, it is paywalled.

https://youtu.be/NHmB7O87ViY
How to post content on Patreon

11:45am. I give up.

https://support.google.com/docs/answer/1663349?hl=en&co=GENIE.Platform%3DDesktop

Let me instead target the Docs themselves.

12pm. I didn't expect that such a serious platform such as Patreon would have such poor text support.

I understand what is wrong.

In Docs, I can set the paragraph spacing to 2x, and it will look the same as in Patreon.

https://bestfriendsclub.ca/ko-fi-vs-patreon/

I am reading this, and it makes it look like there is zero reason for me to pick Patreon over Ko-fi. Patreon just charges higher fees, but offers nothing when it comes to getting more readers.

12:10pm. https://pinepage.com/

I've seen this advertized on Quora.

https://pinepage.com/

I am wondering if I could somehow monetize the gamesoftranscendi blog. Why do the tiers have limits on the number of paying members here.

https://re-library.com/translations/blade-maiden/

This site makes use of Patreon, but it does not host everything there.

https://www.patreon.com/ProjectGB

I wasn't really ready to make my own site.

12:20pm. No. While having the files be on Google Drive is an option, having to manually manage permissions would be a pain in the ass.

Do membership tiers have to be tiered. It would not be bad if I could have separate accessibility...

Yeah, I can actually check the box and select only the Aleator and not the Episteme. That way I could require the members to subscribe to just Heaven's Key or just my blog.

https://www.royalroad.com/forums/thread/122332

Got a reply. He is suggesting that I use PDFs.

Ok, why not. Looking at the amount of achievements these guy has I guess using pdfs hasn't hurt him.

12:35pm. Ok, let me go with that.

12:40pm. Let me get breakfast. I think I understand what I need to do. I'll create two tiers, one for just Simulacrum, one for the Spiral blog priced at 5$/month each.

I'll just use pdfs as that guy suggests. The readers will get used to it at some point. Better for me to do this, than to have to endlessly struggle with formatting.

That is a nice and simple way of doing it.

Having two different accounts would also be hard. Better I keep it all in one place.

...I'll need to make sure to appropriately tag my posts so readers can easily look for them. I'll also make a pinned public post.

12:45pm. Let me have breakfast.

I shouldn't choose Patreon just like that.

https://pinepage.com/

I linked to this before. The mid tier costs 30$ a month which might seem like a lot, but if I had a 100 patrons paying 5$/month each that would cost me 7% a month.

Actually, this site is really cheap if you have a lot of Patrons. 50/month for somebody with 100s of subscribers would be nothing. It is the mid tier that is troublesome.

But why not Ko-fi? Do I have a reason not to pick it over Patreon?

This will take me a while. I should register to Ko-fi and take a look at how it works."