"12:05pm. Seems good. I should release a new version of Spiral. I did a beatiful job on this.
Without a doubt, the way I am doing ref counting now is the way it should be done. I though the other version was solid, but it turns out it had a few bugs. This one is far cleaner and more consistent. Everything is in the right place.

12:15pm. Ok, let me figure out how to publish it, and then I will close the morning session.

...Wait, wait, wait...

What about while statements. By fixing the bug in the if branches could I have cause it to insert it stuff at the end of while statements?

```spiral
            loopw.while (fun _ => true) fun _ =>
                inl _ = b
                ()
            b
```

The reason is no because while branches require unit returns. I am safe!

12:20pm. Let me commit this and then I will publish it. After this I will either do some writing or move on to work on the Python backend.
"