"``` // Does the partial evaluator optimize unused match cases? union t = A : heap i32 * mut (heap u64) | B : heap f64 inl main () =     inl ~a, ~b, ~c = B (heap 5), B (heap 2), B (heap 1)     match (join a,b,c) with     | A, _, _ => 1i32     | _, A, _ => 2     | _, _, A => 3     | _ => 4 ```
I've decided to modify `test39`. I want to see if some of the branches are getting generated correctly.

4:30pm. Have a bunch of unused variable warnings. They don't matter. I am assuming the C compiler will optimize them away.

4:50pm. Done. All tests pass now. Let me push out a new version of Spiral.

4:55pm. I published it on the VS Code marketplace, but why does the NPX command no longer work?

Hmmm...remember how I complained yesterday that cmd doesn't work. It seems that npm doesn't work either. What is happening?

...Uninstalling all those old Python version really wrecked my computer. I need to figure out why npm and cmd commands are not working.

```
PS C:\Users\Marko> where cmd
PS C:\Users\Marko>
```

This is really bizzare.

`cmd` is in the system32 directory, but it is not working.

https://answers.microsoft.com/en-us/windows/forum/all/cmdexe-no-longer-opens-after-uninstalling-python/9ca0dd4a-4eff-4a21-ad30-16d6a8e28790

> Can you check if running "cmd.exe /d" launches the Command Prompt correctly?

Oh this does launch it.

> reg delete "HKCU\SOFTWARE\Microsoft\Command Processor" /v "Autorun" /f

Oh it worked!

I just googled 'window command not working after python uninstall' and it was near the top of the search list. No way I would have figured it out on my own. `npm` is working as well.

> Published mrakgr.spiral-lang-vscode v2.2.6

Whops.

> Unpublishing is not supported at the moment because that might break others who depend on a published extension. We could consider going for an approach like npm, i.e. allow to unpublish only in certain circumstances:

Ah, whatever. It doesn't matter.

5:20pm. My god, how long can it take to publish? At any rate, I am done.

Let me paste what I've written a few entries ago.

///

12:35pm. That's about it. Now I'll have to go though the test cases again and even run them. The purpose of this is to make tail loops faster. I realized during the night that depending on where the jump is made, the compuler could potentially optimize the jump so skips both the decrement and the increment of the reference right at the start of the loop.

For example...

```c
void foo(Heap0 * a) {
    a->refc++;
    ...
    if (cond) {
        a->refc--;
        foo(a);
        }
    }
```

Consider something like this. When the tail call loop optimization is done, instead of jumpting right at the start of the function, it could skip the `a->refc++;` and the `a->refc--;` parts.

When you inline foo once, you see that the inc and dec refs are right next to each other and can be eliminated.

But with the previous function that I had...

```c
void HeapRefc0(Heap0 * x, REFC_FLAG q){
    if (x != NULL) {
        int refc = (x->refc += q & REFC_INCR ? 1 : -1);
        if (!(q & REFC_SUPPR) && refc == 0) {
            HeapRefcBody0(x, REFC_DECR);
            free(x);
        }
    }
}
```

You had things like this. An optimizing compiler is going to have a hard time optimizing this. A particular logical issue that was present here before is that increfing could potentially free the object. Now that is no longer a problem. You just have an increment and a decrement. And they are inlined instead of hidden behind functions.

One consequence is that I can no longer read unitialized objects from arrays, but the assignment to them should be fine. This is perfectly acceptable. The only times I've used nullables is by mistake in F#. Maybe I've used null strings, but that is about it.

...Well, I am doing a search through the project and it does seem like I've used them on occasion. But they are hardly necessary or irreplaceable.

///

This is basically the reason to do it like this. Replacing function calls with simple increment and decrement operations opens the door to optimizing them away in tail calls.

```c
void HeapDecref0(Heap0 * x){
    if (x != NULL && --(x->refc) == 0) { HeapDecrefBody0(x); free(x); }
}
```

Also, I much rather prefer the decrement to be like this. It is very succinct and understandable compared to those boolean operations I had going last time. Remember that it took me time to understand what was going on. I had forgotten that REFC_SUPPR was supposed to be 2 10 in binary and so on. This on the other hand is crystal clear.

I think now I can say that I've done ref counting properly. Maybe in the future I'll extend the heap and mut types so they take in disposal pointers. That would allow them to handle file handle and other kinds of resources instead of just memory. It would be easy to add too.

But it doesn't matter right now. What I should focus next is dealing with the Python backend. Once I have it, I can begin work on the UPMEM composite backend. After I have that I can begin work on the generic map, fold and filter kernels. And once I have that I'll be able to write an article for HN.

I'll also do a C based UPMEM backend as well.

I am not sure if doing this will benefit me much, but it will be a good showcase of my skills."