"I found an interesting email in the spam filter. I'll reply to it tomorrow.
https://news.ycombinator.com/item?id=34411009
Tell HN: Gmail's spam filters have gone bonkers

I only noticed it because there was a single email there, and I only peeked because I saw this thread on HN a while ago.

///

> P.S. Also looking for a good alternative to gmail if anybody has some suggestions.
100% Fastmail. I've been on them for eight years now, after switching from a decade of Gmail. I've had zero complaints.

Switching from Gmail isn't even that bad - you can keep your Google account using your new email address as the ID, so that you can still use Docs and Drive and such. I wrote a short guide here: https://www.justus.ws/tech/how-to-ditch-gmail/

///

I'll keep this in mind for the future.

> I switched to fastmail about a year ago, and I still have my gmail piping into it as an external mailbox. Fastmail's spam filter consistently catches gmail's false negatives.

Hmmm, really?

1/18/2023

8:45am. First of all, let me fix a few bugs I realized the constraint process still has.

```fs
            match b with
            | TyMetavar(x,_) -> x.constraints <- Set.union con x.constraints; []
```

This thing here does not go through a kind check.

```fs
| con, TyVar v & x -> if Set.contains con v.constraints then [] else [ConstraintError(con,x)]
```

I'll also get rid of this case as it is redundant.

```fs
        let rec constraint_process (con : Constraint Set) b =
            let unify_kind got expected = unify_kind' KindErrorInConstraint r got expected
            let body = function
                | CUInt, TyPrim (UInt8T | UInt16T | UInt32T | UInt64T)
                | CSInt, TyPrim (Int8T | Int16T | Int32T | Int64T)
                | CInt, TyPrim (UInt8T | UInt16T | UInt32T | UInt64T | Int8T | Int16T | Int32T | Int64T)
                | CFloat, TyPrim (Float32T | Float64T)
                | CNumber, TyPrim (UInt8T | UInt16T | UInt32T | UInt64T | Int8T | Int16T | Int32T | Int64T | Float32T | Float64T)
                | CPrim, TyPrim _
                | CSymbol, TySymbol _
                | CRecord, TyRecord _ -> []
                | con, TyMetavar(x,_) -> x.constraints <- Set.add con x.constraints; []
                | CPrototype prot & con, x ->
                    match type_apply_split x with
                    | TyNominal ins, x' ->
                        match Map.tryFind (prot,ins) top_env.prototypes_instances with
                        | Some cons ->
                            try List.fold2 (fun ers con x -> List.append (constraint_process con (visit_t x)) ers) [] cons x'
                            with :? System.ArgumentException -> [] // This case can occur due when kind application overflows in a previous expression.
                        | None -> [InstanceNotFound(prot,ins)]
                    | TyMetavar _ & x, _ -> [PrototypeConstraintCannotPropagateToMetavar(prot,x)]
                    | TyVar _ & x, _ -> [PrototypeConstraintCannotPropagateToVar(prot,x)]
                    | _ -> [ConstraintError(con,x)]
                | con, x -> [ConstraintError(con,x)]

            match b with
            | TyVar b -> if con.IsSubsetOf b.constraints = false then [ForallVarConstraintError(b.name,con,b.constraints)] else []
            | b ->
                let b_kind = tt top_env b
                Set.fold (fun ers con ->
                    unify_kind b_kind (constraint_kind top_env con)
                    List.append (body (con,b)) ers
                    ) [] con
```

The bug I fixed today is that when propagating nested contraints to metavars, the kind check would be erroneously short circuited.

I neglected to write about what the purpose of the rewrite yesterday was, but yesterday since the contraints did not have a kind, they could be of any kind at all. This made it very confusing to me to understand what the rules of propagating nested contraint should be. And indeed, what I had here two days ago made no sense. Previously I had that unapplied forall check, then I removed it. I simply did not understand that contraints should have kinds. Contraints are like the type of types.

```fs
                    | TyMetavar _ & x, _ -> [PrototypeConstraintCannotPropagateToMetavar(prot,x)]
                    | TyVar _ & x, _ -> [PrototypeConstraintCannotPropagateToVar(prot,x)]
```

Also this part requires some comment. I spent a lot of time thinking about this rule. If I didn't have the metavar one for example, and it got unified to a principal type, then it could be that reruning the check here with a nominal would have resulted in an error.

I thought of how to deal with this and I realized that I would need to collect all the contraints and resolve them after the regular unification is done. This would be a difficult engineering problem that I did not feel like doing, so I opted to just require type annotations here. This saves me a lot of work.

```fs
| PrototypeConstraintCannotPropagateToMetavar(a,b) -> sprintf "Cannot propagate the %s prototype constraint to the applied metavariable as the kinds would not match. If this is not intended to be a type var, provide a type annotation to a concrete type.\nGot: %s" env.prototypes.[a].name (f b)
```

I'll modify this case so the error is more informative.

Let me test this a little and then I will publish a new version. Then it is bath time.

9:05am. It seems like it works, I won't do extensive testing.

9:10am. Published v2.3.10. This should be the definite version of Spiral.

Now, bath time. I'll think of what I want to do today after that. Probably reply to that email I found in the spam folder yesterday."